+++ tm-worker/src/main/java/com/tm/worker/core/node/function/encoder/EncodeURIComponentNode.java
@@ -6,8 +6,6 @@ import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
-import java.net.URLEncoder;
-
 
 @Slf4j
 public class EncodeURIComponentNode extends FunctionNode {
