+++ tm-worker/src/main/java/com/tm/worker/core/protocol/http/HttpClientBuilderProxy.java
@@ -1,7 +1,10 @@
 package com.tm.worker.core.protocol.http;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.http.conn.ssl.*;
+import org.apache.http.conn.ssl.DefaultHostnameVerifier;
+import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
+import org.apache.http.conn.ssl.TrustAllStrategy;
+import org.apache.http.conn.ssl.TrustStrategy;
 import org.apache.http.impl.client.BasicCookieStore;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
