+++ tm-worker/src/main/java/com/tm/worker/core/threads/AutoTestContext.java
@@ -1,9 +1,9 @@
 package com.tm.worker.core.threads;
 
 import com.tm.worker.core.cookie.AutoTestCookie;
-import com.tm.worker.core.task.TaskService;
 import com.tm.worker.core.task.CaseTask;
 import com.tm.worker.core.task.PlanTask;
+import com.tm.worker.core.task.TaskService;
 import com.tm.worker.core.variable.AutoTestVariables;
 import lombok.Data;
